refractor code

+++ src/execution/delete_executor.cpp
@@ -18,14 +18,12 @@ namespace bustub {
 
 DeleteExecutor::DeleteExecutor(ExecutorContext *exec_ctx, const DeletePlanNode *plan,
                                std::unique_ptr<AbstractExecutor> &&child_executor)
-    : AbstractExecutor(exec_ctx), plan_(plan), child_executor_(std::move(child_executor)) {}
-
-void DeleteExecutor::Init() {
+    : AbstractExecutor(exec_ctx), plan_(plan), child_executor_(std::move(child_executor)) {
   catalog_ = exec_ctx_->GetCatalog();
   tb_info_ = catalog_->GetTable(plan_->TableOid());
-  tb_hp_ = tb_info_->table_.get();
-  child_executor_->Init();
 }
+
+void DeleteExecutor::Init() { child_executor_->Init(); }
 
 bool DeleteExecutor::Next([[maybe_unused]] Tuple *tuple, RID *rid) {
   LockManager *lock_mgr = exec_ctx_->GetLockManager();
@@ -35,46 +33,36 @@ bool DeleteExecutor::Next([[maybe_unused]] Tuple *tuple, RID *rid) {
   Tuple old_tuple;
   RID old_rid;
   while (true) {
-    try {
-      if (child_executor_->Next(&old_tuple, &old_rid)) {
-        break;
-      }
-    } catch (Exception &e) {
-      throw Exception(ExceptionType::UNKNOWN_TYPE, "DeleteError:child execute error.");
-      return false;
+    if (child_executor_->Next(&old_tuple, &old_rid)) {
+      break;
     }
 
     // add lock
     if (txn->GetIsolationLevel() != IsolationLevel::REPEATABLE_READ) {
-      if (!lock_mgr->LockExclusive(txn, *rid)) {
+      if (!lock_mgr->LockExclusive(txn, old_rid)) {
         txn_mgr->Abort(txn);
       }
     } else {
-      if (!lock_mgr->LockUpgrade(txn, *rid)) {
+      if (!lock_mgr->LockUpgrade(txn, old_rid)) {
         txn_mgr->Abort(txn);
       }
     }
 
     // delete
-    tb_hp_->MarkDelete(old_rid, exec_ctx_->GetTransaction());
+    if (tb_info_->table_->MarkDelete(old_rid, exec_ctx_->GetTransaction())) {
+      // delete index
+      auto idxinfo_arr = catalog_->GetTableIndexes(tb_info_->name_);
 
-    // delete index
-    auto idxinfo_arr = catalog_->GetTableIndexes(tb_info_->name_);
+      for (auto &idxinfo : idxinfo_arr) {
+        idxinfo->index_->DeleteEntry(old_tuple.KeyFromTuple(*child_executor_->GetOutputSchema(), idxinfo->key_schema_,
+                                                            idxinfo->index_->GetKeyAttrs()),
+                                     old_rid, exec_ctx_->GetTransaction());
+        // record the old tuple for rollback
+        IndexWriteRecord iw_record(old_rid, tb_info_->oid_, WType::DELETE, old_tuple, old_tuple, idxinfo->index_oid_,
+                                   exec_ctx_->GetCatalog());
 
-    for (auto &idxinfo : idxinfo_arr) {
-      idxinfo->index_->DeleteEntry(old_tuple.KeyFromTuple(*child_executor_->GetOutputSchema(), idxinfo->key_schema_,
-                                                          idxinfo->index_->GetKeyAttrs()),
-                                   old_rid, exec_ctx_->GetTransaction());
-      // record the old tuple for rollback
-      IndexWriteRecord iw_record(old_rid, tb_info_->oid_, WType::DELETE, old_tuple, old_tuple, idxinfo->index_oid_,
-                                 exec_ctx_->GetCatalog());
-
-      txn->GetIndexWriteSet()->emplace_back(iw_record);
-    }
-
-    // unlock
-    if (lock_mgr != nullptr && txn->GetIsolationLevel() == IsolationLevel::READ_COMMITTED) {
-      lock_mgr->Unlock(txn, old_rid);
+        txn->GetIndexWriteSet()->emplace_back(iw_record);
+      }
     }
   }
